requirements.txt
fastapi==0.95.2
langchain==0.0.240
pydantic==1.10.12
python-dotenv==1.0.0
SQLAlchemy==2.0.19
uvicorn==0.23.2
psycopg2-binary==2.9.9
openai==0.27.8
httpx==0.25.0